+++ examples/slack-connect-example/hooks/index.ts
@@ -1,1 +1,1 @@
-export {default as useSetToken} from "./useSetToken"
+export { default as useSetToken } from "./useSetToken";