+++ packages/client/.eslintrc.js
@@ -2,4 +2,7 @@
 module.exports = {
   root: true,
   extends: ["@knocklabs/eslint-config/library.js"],
+  parserOptions: {
+    projects: ["tsconfig.json", "tsconfig.node.json"],
+  },
 };