+++ packages/react-native/vite.config.ts
@@ -1,6 +1,6 @@
-import { LibraryFormats, defineConfig, loadEnv } from "vite";
-import { resolve } from "path";
 import react from "@vitejs/plugin-react";
+import { resolve } from "path";
+import { LibraryFormats, defineConfig, loadEnv } from "vite";
 import dts from "vite-plugin-dts";
 import noBundlePlugin from "vite-plugin-no-bundle";