+++ packages/react/.eslintrc.js
@@ -3,10 +3,8 @@ module.exports = {
   root: true,
   extends: [
     "@knocklabs/eslint-config/library.js",
-    "plugin:@typescript-eslint/recommended",
     "plugin:react-hooks/recommended",
   ],
-  parser: "@typescript-eslint/parser",
   parserOptions: {
     projects: ["tsconfig.json", "tsconfig.node.json"],
   },