+++ packages/react/src/modules/slack/components/SlackAddChannelInput/index.ts
@@ -1,1 +1,1 @@
-export * from "./SlackAddChannelInput"
+export * from "./SlackAddChannelInput";