+++ packages/react/src/modules/slack/components/SlackAuthButton/index.ts
@@ -1,1 +1,1 @@
-export * from "./SlackAuthButton"
+export * from "./SlackAuthButton";