+++ packages/client/CHANGELOG.md
@@ -1,5 +1,19 @@
 # Changelog
 
+## 0.10.11
+
+### Patch Changes
+
+- 395f0ca: fix: check type of zustand default import and fix cjs build
+
+## 0.10.10
+
+### Patch Changes
+
+- a4d520c: chore: update generic types
+- Updated dependencies [a4d520c]
+  - @knocklabs/types@0.1.4
+
 ## 0.10.9
 
 ### Patch Changes