+++ packages/react-native/CHANGELOG.md
@@ -1,5 +1,22 @@
 # Changelog
 
+## 0.2.7
+
+### Patch Changes
+
+- Updated dependencies [395f0ca]
+  - @knocklabs/client@0.10.11
+  - @knocklabs/react-core@0.2.22
+
+## 0.2.6
+
+### Patch Changes
+
+- a4d520c: chore: update generic types
+- Updated dependencies [a4d520c]
+  - @knocklabs/react-core@0.2.21
+  - @knocklabs/client@0.10.10
+
 ## 0.2.5
 
 ### Patch Changes